+++ CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `from_gateway_bot` classmethods can now also take cache-less `ShardAware` bots.
 - Bumped minimum Hikari version to `2.0.0.dev116`.
+- Renamed `ErrorManager.with_rule` to [ErrorManager.add_rule][yuyo.backoff.ErrorManager.add_rule]
+  and make its arguments positional only.
+- The `guild` argument for [yuyo.chunk_tracker.ChunkTracker.request_guild_members][] is now
+  positional only.
+- `iterator` is now positional only in
+  [ReactionPaginator.\_\_init\_\_][yuyo.reactions.ReactionPaginator.__init__] and
+  [ComponentPaginator.\_\_init\_\_][yuyo.components.ComponentPaginator.__init__].
+- `lines` is now positional only in [yuyo.pagination.async_paginate_string][],
+  [yuyo.pagination.sync_paginate_string][] and [yuyo.pagination.paginate_string][].
 
 ### Fixed
 - `Context.create_initial_response` (and by extension `Context.respond` for the initial